% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsets.R
\name{as.data.frame.fsets}
\alias{as.data.frame.fsets}
\alias{as.matrix.fsets}
\title{Convert an object of \code{fsets} class into a matrix or data frame
This function converts an instance of S3 class \link{fsets} into a
matrix or a data frame. The \code{\link[=vars]{vars()}} and \code{\link[=specs]{specs()}} attributes
of the original object are deleted.}
\usage{
\method{as.data.frame}{fsets}(x, ...)

\method{as.matrix}{fsets}(x, ...)
}
\arguments{
\item{x}{An instance of class \link{fsets} to be converted}

\item{...}{arguments further passed to \code{as.data.frame} after converting
to matrix in \code{as.data.frame.fsets}. Unused in \code{as.matrix.fsets}.}
}
\value{
A numeric matrix or data frame of membership degrees.
}
\description{
Convert an object of \code{fsets} class into a matrix or data frame
This function converts an instance of S3 class \link{fsets} into a
matrix or a data frame. The \code{\link[=vars]{vars()}} and \code{\link[=specs]{specs()}} attributes
of the original object are deleted.
}
\seealso{
\code{\link[=fsets]{fsets()}}, \code{\link[=fcut]{fcut()}}, \code{\link[=lcut]{lcut()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
