% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsets.R
\name{cbind.fsets}
\alias{cbind.fsets}
\title{Combine several 'fsets' objects into a single one}
\usage{
\method{cbind}{fsets}(..., deparse.level = 1)
}
\arguments{
\item{...}{A sequence of objects of class 'fsets' to be merged by columns.}

\item{deparse.level}{This argument has currently no function and is added
here only for compatibility with generic \code{\link[=cbind]{cbind()}} function.}
}
\value{
An object of class 'fsets' that is created by merging the arguments
by columns.  Also the arguments' attributes \code{\link[=vars]{vars()}} and \code{\link[=specs]{specs()}} are merged together.
}
\description{
Take a sequence of objects of class 'fsets' and combine them by columns.
This version of cbind takes care of the \code{\link[=vars]{vars()}} and \code{\link[=specs]{specs()}}
attributes of the arguments and merges them to the result. If some argument
does not inherit from the 'fsets' class, an error is thrown.
}
\details{
The \code{\link[=vars]{vars()}} attribute is merged by concatenating the \code{\link[=vars]{vars()}} attributes
of each argument. Also the \code{\link[=specs]{specs()}} attributes of the arguments are merged together.
}
\seealso{
\code{\link[=vars]{vars()}}, \code{\link[=specs]{specs()}}, \code{\link[=fcut]{fcut()}}, \code{\link[=lcut]{lcut()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
