% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedge.R
\name{defaultHedgeParams}
\alias{defaultHedgeParams}
\title{A list of the parameters that define the shape of the hedges.}
\usage{
defaultHedgeParams
}
\description{
A list of the parameters that define the shape of the hedges.
}
