% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farules.R
\name{is.farules}
\alias{is.farules}
\title{Test whether \code{x} inherits from the S3 \code{farules} class.}
\usage{
is.farules(x)
}
\arguments{
\item{x}{An object being tested.}
}
\value{
\code{TRUE} if \code{x} is a valid \code{\link[=farules]{farules()}} object and \code{FALSE} otherwise.
}
\description{
Test whether \code{x} inherits from the S3 \code{farules} class.
}
\seealso{
\code{\link{farules}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
