% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{is.ft}
\alias{is.ft}
\title{Test whether \code{x} is a valid object of the S3 \code{ft} class}
\usage{
is.ft(x)
}
\arguments{
\item{x}{An object to be tested}
}
\value{
\code{TRUE} if \code{x} is a valid \code{ft} object and \code{FALSE} otherwise.
}
\description{
Test whether \code{x} has a valid format for objects of the S3 \code{ft} class that represents
the Fuzzy Transform.
}
\details{
This function tests whether \code{x} is an instance of the \code{ft} class and whether it is a list
with the following elements: \code{inputs} character vector, \code{partitions} list, \code{order} number,
\code{antecedents} matrix and \code{consequents} matrix.
}
\seealso{
\code{\link[=ft]{ft()}}, \code{\link[=ftinv]{ftinv()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
