% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfl-package.R
\docType{package}
\name{lfl}
\alias{lfl}
\title{lfl - Linguistic Fuzzy Logic}
\description{
Various algorithms related to linguistic fuzzy logic: mining for linguistic fuzzy association
rules, composition of fuzzy relations, performing perception-based logical deduction (PbLD),
and forecasting time-series using fuzzy rule-based ensemble (FRBE).
}
