% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mase.R
\name{mase}
\alias{mase}
\title{Compute Mean Absolute Scaled Error (MASE)}
\usage{
mase(forecast, validation)
}
\arguments{
\item{forecast}{A numeric vector of forecasted values}

\item{validation}{A numeric vector of actual (real) values}
}
\value{
A Mean Absolute Scaled Error (MASE)
}
\description{
MASE is computed as \eqn{sum(abs(validation - forecast)) / sum(abs(validation[-1] - validation[-n])) / (n/(n-1))}.
}
\seealso{
\code{\link[=rmse]{rmse()}}, \code{\link[=smape]{smape()}}, \code{\link[=frbe]{frbe()}}
}
\author{
Michal Burda
}
