% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{print.algebra}
\alias{print.algebra}
\title{Print an instance of the \code{\link[=algebra]{algebra()}} S3 class in a human readable form.}
\usage{
\method{print}{algebra}(x, ...)
}
\arguments{
\item{x}{An instance of the \code{\link[=algebra]{algebra()}} S3 class}

\item{...}{Unused.}
}
\value{
None.
}
\description{
Print an instance of the \code{\link[=algebra]{algebra()}} S3 class in a human readable form.
}
\seealso{
\code{\link[=algebra]{algebra()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
