% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ctx.R
\name{print.ctx3}
\alias{print.ctx3}
\alias{print.ctx5}
\alias{print.ctx3bilat}
\alias{print.ctx5bilat}
\title{Print the linguistic context}
\usage{
\method{print}{ctx3}(x, ...)

\method{print}{ctx5}(x, ...)

\method{print}{ctx3bilat}(x, ...)

\method{print}{ctx5bilat}(x, ...)
}
\arguments{
\item{x}{A linguistic context to be printed}

\item{...}{Unused.}
}
\value{
Nothing.
}
\description{
Format an object of the \code{\link[=ctx3]{ctx3()}}, \code{\link[=ctx5]{ctx5()}}, \code{\link[=ctx3bilat]{ctx3bilat()}} and the \code{\link[=ctx5bilat]{ctx5bilat()}}
class into human readable form and print it to the output.
}
\examples{

  context <- ctx3()
  print(context)

}
\seealso{
\code{\link[=ctx3]{ctx3()}}, \code{\link[=ctx5]{ctx5()}}, \code{\link[=ctx3bilat]{ctx3bilat()}}, \code{\link[=ctx5bilat]{ctx5bilat()}}, \code{\link[=minmax]{minmax()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
