% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farules.R
\name{print.farules}
\alias{print.farules}
\title{Print an instance of the \code{\link[=farules]{farules()}} S3 class in a human readable form.}
\usage{
\method{print}{farules}(x, ...)
}
\arguments{
\item{x}{An instance of the \code{\link[=farules]{farules()}} S3 class}

\item{...}{Unused.}
}
\value{
None.
}
\description{
Print an instance of the \code{\link[=farules]{farules()}} S3 class in a human readable form.
}
\seealso{
\code{\link[=farules]{farules()}}, \code{\link[=searchrules]{searchrules()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
