% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsets.R
\name{print.fsets}
\alias{print.fsets}
\title{Print an instance of the \code{\link[=fsets]{fsets()}} class}
\usage{
\method{print}{fsets}(x, ...)
}
\arguments{
\item{x}{An instance of the \code{\link[=fsets]{fsets()}} class}

\item{...}{Unused.}
}
\value{
Nothing
}
\description{
Format an object of the \code{\link[=fsets]{fsets()}} class into human readable form
and print it to the output.
}
\examples{
    d <- fcut(CO2[, 1:2])
    print(d)

}
\seealso{
\code{\link[=fsets]{fsets()}}, \code{\link[=fcut]{fcut()}}, \code{\link[=lcut]{lcut()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
