% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Compute Root Mean Squared Error (RMSE)}
\usage{
rmse(forecast, validation)
}
\arguments{
\item{forecast}{A numeric vector of forecasted values}

\item{validation}{A numeric vector of actual (real) values}
}
\value{
A Root Mean Squared Error (RMSE)
}
\description{
RMSE is computed as \eqn{sqrt(mean((forecast - validation)^2))}.
}
\seealso{
\code{\link[=smape]{smape()}}, \code{\link[=mase]{mase()}}, \code{\link[=frbe]{frbe()}}
}
\author{
Michal Burda
}
