\name{slices}
\alias{slices}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Return vector of values from given interval}
\description{
    Returns an ordered vector of values from given interval, of given size, generated by equal
    steps.
}
\usage{
slices(from, to, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
    \item{from}{
        The lower bound of the interval.
    }
    \item{to}{
        The upper bound of the interval.
    }
    \item{n}{
        The length of the vector to be produced.
    }
}
\details{
    Returns a vector of values from \code{from} to \code{to} (inclusive), with equal difference
    between two consecutive values, with total length \code{n}. Function is useful e.g. together
    with the \code{\link{pbld}} or \code{\link{defuzz}} functions (for the \code{values}
    argument; see also \code{\link{lcut}} or \code{\link{fcut}}) or \code{\link{defuzz}}).
}
\value{
    A vector of numbers in the given interval and size.
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
    Michal Burda
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
    \code{\link{pbld}},
    \code{\link{defuzz}},
    \code{\link{fcut}},
    \code{\link{lcut}}
}
\examples{
    slices(1, 5, 10) # 1, 1.5, 2, 2.5, 3, 3.5 4, 4.5, 5

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }
\keyword{ robust }
