% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsets.R
\name{str.fsets}
\alias{str.fsets}
\title{Compactly display the internal structure of an instance of the \code{\link[=fsets]{fsets()}} class}
\usage{
\method{str}{fsets}(object, ...)
}
\arguments{
\item{object}{An instance of the \code{\link[=fsets]{fsets()}} class}

\item{...}{Unused.}
}
\value{
Nothing
}
\description{
This function is a method for the generic function \code{\link[=str]{str()}} for objects of the
\code{\link[=fsets]{fsets()}} class. It is a compact way to display the internal structure of an
instance of the \code{\link[=fsets]{fsets()}} class.
}
\seealso{
\code{\link[=fsets]{fsets()}}, \code{\link[=fcut]{fcut()}}, \code{\link[=lcut]{lcut()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
