% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs.R
\name{triangle}
\alias{triangle}
\alias{raisedcos}
\alias{raisedcos}
\title{Factories for functions that convert numeric data into membership degrees of fuzzy sets}
\usage{
triangle(lo, center, hi)

raisedcos(lo, center, hi)
}
\arguments{
\item{lo}{A lower bound (can be -Inf).}

\item{center}{A peak value.}

\item{hi}{An upper bound (can be Inf).}
}
\value{
A function with single argument \code{x} that should be a numeric vector to be converted.
}
\description{
These functions create functions with a single argument \code{x} that compute membership degrees of \code{x} to a fuzzy set
of either triangular or raised-cosine shape that is defined by \code{lo}, \code{center}, and \code{hi}.
}
\details{
The arguments must satisfy \code{lo \le center \le hi}. Functions compute membership degrees of triangular or
raised-cosine fuzzy sets. \code{x} values equal to \code{center obtain membership degree equal to 1,}x\code{values lower than}lo\code{or greater than}hi\code{obtain memberhsip degree equal to 0. A transition of the triangular (resp. raised cosinal) shape (with peak at}center\code{is computed for}x\code{values between}lo\code{and}hi`.

If \code{lo == -Inf} then any value that is lower or equal to center gets memberhsip degree 1.  Similarly, if \code{hi == Inf}
then any value that is greater or equal to center gets memberhsip degree 1.

\code{triangle} produces fuzzy sets of a triangular shape (with peak at \code{center}), \code{raisedcos} produces
fuzzy sets defined as a raised cosine hill.
}
\examples{

plot(triangle(-1, 0, 1), from=-2, to=3)
plot(triangle(-1, 0, 2), from=-2, to=3)
plot(triangle(-Inf, 0, 1), from=-2, to=3)
plot(triangle(-1, 0, Inf), from=-2, to=3)

plot(raisedcos(-1, 0, 1), from=-2, to=3)
plot(raisedcos(-1, 0, 2), from=-2, to=3)
plot(raisedcos(-Inf, 0, 1), from=-2, to=3)
plot(raisedcos(-1, 0, Inf), from=-2, to=3)

}
\seealso{
\code{\link{fcut}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{multivariate}
\keyword{robust}
