% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farules.R
\name{as.data.frame.farules}
\alias{as.data.frame.farules}
\title{Convert the instance of the \code{\link[=farules]{farules()}} S3 class into a data frame.
Empty \code{\link[=farules]{farules()}} object is converted into an empty \code{\link[=data.frame]{data.frame()}}.}
\usage{
\method{as.data.frame}{farules}(x, ...)
}
\arguments{
\item{x}{An instance of class \code{\link[=farules]{farules()}} to be transformed.}

\item{...}{Unused.}
}
\value{
A data frame of statistics of the rules that are stored in the given
\code{\link[=farules]{farules()}} object. Row names of the resulting data frame are in
the form: \verb{A1 & A2 & ... & An => C}, where \code{Ai} are antecedent
predicates and \code{C} is a consequent. An empty \code{\link[=farules]{farules()}} object
is converted into an empty \code{\link[=data.frame]{data.frame()}} object.
}
\description{
Convert the instance of the \code{\link[=farules]{farules()}} S3 class into a data frame.
Empty \code{\link[=farules]{farules()}} object is converted into an empty \code{\link[=data.frame]{data.frame()}}.
}
\seealso{
\code{\link[=farules]{farules()}}, \code{\link[=searchrules]{searchrules()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{multivariate}
\keyword{robust}
