% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedge.R
\docType{data}
\name{defaultHedgeParams}
\alias{defaultHedgeParams}
\title{A list of the parameters that define the shape of the hedges.}
\format{
An object of class \code{list} of length 9.
}
\usage{
defaultHedgeParams
}
\description{
A list of the parameters that define the shape of the hedges.
}
\keyword{datasets}
