% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsets.R
\name{is.fsets}
\alias{is.fsets}
\title{Test whether \code{x} is a valid object of the S3 \code{fsets} class}
\usage{
is.fsets(x)
}
\arguments{
\item{x}{An object being tested.}
}
\value{
TRUE if \code{x} is a valid \code{fsets} object and FALSE otherwise.
}
\description{
This function tests wheter \code{x} inherits from S3 \code{fsets} class.
}
\seealso{
\code{\link[=fsets]{fsets()}}
}
\author{
Michal Burda
}
\keyword{models}
\keyword{robust}
