% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfl-package.R
\docType{package}
\name{lfl}
\alias{lfl}
\title{lfl - Linguistic Fuzzy Logic}
\description{
Various algorithms related (not only) to linguistic fuzzy logic: mining for linguistic fuzzy association
rules, composition of fuzzy relations, performing perception-based logical deduction (PbLD),
and forecasting time-series using fuzzy rule-based ensemble (FRBE).
}
\details{
Other methods include fuzzy transform and computation of Sugeno integrals. Also basic algebraic
functions related to fuzzy logic are contained, which allow to handle missing values using different
styles such as Kleene, Bochvar, Sobocinski and other.
}
