\name{print.farules}
\alias{print.farules}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print an instance of the \code{\link{farules}} class}
\description{
    Format an object of the \code{\link{farules}} class into human readable form and print it to the
    output.
}
\usage{
\method{print}{farules}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
    \item{x}{
        An instance of \code{\link{farules}} class
    }
    \item{...}{
        Unused.
    }
}
\details{
    Format an object of the \code{\link{farules}} class into human readable form and print it to the
    output.
}
\value{
    None.
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
    Michal Burda
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
    \code{\link{farules}},
    \code{\link{searchrules}},
    \code{\link{reduce}},
    \code{\link{sel}}
}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }
\keyword{ robust }
