% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smape.R
\name{smape}
\alias{smape}
\title{Compute Symmetric Mean Absolute Percentage Error (SMAPE)}
\usage{
smape(forecast, validation)
}
\arguments{
\item{forecast}{A numeric vector of forecasted values}

\item{validation}{A numeric vector of actual (real) values}
}
\value{
A Symmetric Mean Absolute Percentage Error (SMAPE)
}
\description{
SMAPE is computed as \eqn{mean(abs(forecast - validation) / ((abs(forecast) + abs(validation)) / 2))}.
}
\seealso{
\code{\link[=rmse]{rmse()}}, \code{\link[=mase]{mase()}}, \code{\link[=frbe]{frbe()}}
}
\author{
Michal Burda
}
