% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprsDeprecated.R
\name{triangle}
\alias{triangle}
\alias{raisedcos}
\title{Deprecated functions to compute membership degrees of numeric fuzzy sets}
\usage{
triangle(x, lo, center, hi)

raisedcos(x, lo, center, hi)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{lo}{A lower bound (can be -Inf).}

\item{center}{A peak value.}

\item{hi}{An upper bound (can be Inf).}
}
\value{
A numeric vector of membership degrees of \code{x} to a fuzzy set with the shape
determined with \code{lo}, \code{center}, \code{hi}.
}
\description{
These functions compute membership degrees of numeric fuzzy sets with
triangular or raised-cosine shape. These functions are \emph{deprecated}.
Please use \code{\link[=triangular]{triangular()}} or \code{\link[=raisedcosine]{raisedcosine()}} functions instead.
}
\seealso{
\code{\link[=triangular]{triangular()}}, \code{\link[=raisedcosine]{raisedcosine()}}
}
\author{
Michal Burda
}
